% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglinedensity-package.R
\docType{package}
\name{gglinedensity-package}
\alias{gglinedensity-package}
\title{gglinedensity: Make DenseLines Heatmaps with 'ggplot2'}
\description{
Visualise overlapping time series lines as a heatmap of line density. Provides a 'ggplot2' statistic implementing the DenseLines algorithm, which "normalizes time series by the arc length to compute accurate densities" (Moritz and Fisher, 2018) \doi{10.48550/arXiv.1808.06019}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrryt/gglinedensity}
  \item \url{https://hrryt.github.io/gglinedensity/}
  \item Report bugs at \url{https://github.com/hrryt/gglinedensity/issues}
}

}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk} [copyright holder]

Authors:
\itemize{
  \item Dominik Moritz \email{domoritz@gmail.com} (Rust original, <https://github.com/domoritz/line-density-rust>)
}

Other contributors:
\itemize{
  \item The authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
  \item Hiroaki Yutani \email{yutani.ini@gmail.com} (\href{https://orcid.org/0000-0002-3385-7233}{ORCID}) (Scripts where noted, <https://yutannihilation.github.io/string2path>) [contributor]
}

}
\keyword{internal}
